services/api-gateway/requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
asyncpg==0.29.0
sqlalchemy==2.0.31
alembic==1.13.2
psycopg2-binary==2.9.9

# Authentication & Security
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.8
pyjwt==2.9.0
pyotp==2.9.0
qrcode[pil]==7.4.2

# Rate Limiting
slowapi==0.1.9
redis==5.0.1
hiredis==2.3.2

# Email
sendgrid==6.11.0
jinja2==3.1.4

# SMS (Twilio)
twilio==9.2.3

# SSO
python3-saml==1.16.0
authlib==1.3.1

# Validation
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0

# HTTP Client
httpx==0.27.0
aiohttp==3.10.0

# Monitoring
prometheus-client==0.20.0
sentry-sdk[fastapi]==2.12.0

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0
ua-parser==0.18.0
maxminddb==2.6.2

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
httpx==0.27.0
freezegun==1.5.1
